Added a global UserRole context that can provide the user's role from any page
